Move all analyzers and producers to plugins directory
