+++ Validation/MuonGEMDigis/plugins/SealModule.cc
@@ -1,0 +1,14 @@
+#include "Validation/MuonGEMDigis/plugins/GEMStripDigiValidation.h"
+#include "Validation/MuonGEMDigis/plugins/GEMPadDigiValidation.h"
+#include "Validation/MuonGEMDigis/plugins/GEMPadDigiClusterValidation.h"
+#include "Validation/MuonGEMDigis/plugins/GEMCoPadDigiValidation.h"
+#include "Validation/MuonGEMDigis/plugins/GEMDigiTrackMatch.h"
+#include "Validation/MuonGEMDigis/plugins/GEMCheckGeometry.h"
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(GEMStripDigiValidation);
+DEFINE_FWK_MODULE(GEMPadDigiValidation);
+DEFINE_FWK_MODULE(GEMPadDigiClusterValidation);
+DEFINE_FWK_MODULE(GEMCoPadDigiValidation);
+DEFINE_FWK_MODULE(GEMDigiTrackMatch);
+DEFINE_FWK_MODULE(GEMCheckGeometry);
